+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_template_builder.py
@@ -243,18 +243,19 @@ def build_msi_role_assignment(vm_vmss_name, vm_vmss_resource_id, role_definition
 
 def build_vm_resource(  # pylint: disable=too-many-locals
         cmd, name, location, tags, size, storage_profile, nics, admin_username,
-        availability_set_id=None, admin_password=None, ssh_key_value=None, ssh_key_path=None,
+        availability_set_id=None, admin_password=None, ssh_key_values=None, ssh_key_path=None,
         image_reference=None, os_disk_name=None, custom_image_os_type=None, authentication_type=None,
         os_publisher=None, os_offer=None, os_sku=None, os_version=None, os_vhd_uri=None,
         attach_os_disk=None, os_disk_size_gb=None, custom_data=None, secrets=None, license_type=None, zone=None,
-        disk_info=None, boot_diagnostics_storage_uri=None, ultra_ssd_enabled=None, proximity_placement_group=None):
+        disk_info=None, boot_diagnostics_storage_uri=None, ultra_ssd_enabled=None, proximity_placement_group=None,
+        computer_name=None):
 
     os_caching = disk_info['os'].get('caching')
 
     def _build_os_profile():
 
         os_profile = {
-            'computerName': name,
+            'computerName': computer_name or name,
             'adminUsername': admin_username
         }
 
@@ -264,7 +265,7 @@ def build_vm_resource(  # pylint: disable=too-many-locals
         if custom_data:
             os_profile['customData'] = b64encode(custom_data)
 
-        if ssh_key_value and ssh_key_path:
+        if ssh_key_values and ssh_key_path:
             os_profile['linuxConfiguration'] = {
                 'disablePasswordAuthentication': authentication_type == 'ssh',
                 'ssh': {
@@ -272,7 +273,7 @@ def build_vm_resource(  # pylint: disable=too-many-locals
                         {
                             'keyData': ssh_key_value,
                             'path': ssh_key_path
-                        }
+                        } for ssh_key_value in ssh_key_values
                     ]
                 }
             }
@@ -614,7 +615,7 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
                         vm_sku, instance_count, ip_config_name, nic_name, subnet_id,
                         public_ip_per_vm, vm_domain_name, dns_servers, nsg, accelerated_networking,
                         admin_username, authentication_type, storage_profile, os_disk_name, disk_info,
-                        os_type, image=None, admin_password=None, ssh_key_value=None,
+                        os_type, image=None, admin_password=None, ssh_key_values=None,
                         ssh_key_path=None, os_publisher=None, os_offer=None, os_sku=None, os_version=None,
                         backend_address_pool_id=None, inbound_nat_pool_id=None, health_probe=None,
                         single_placement_group=None, platform_fault_domain_count=None, custom_data=None,
@@ -710,7 +711,7 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
     if admin_password:
         os_profile['adminPassword'] = "[parameters('adminPassword')]"
 
-    if ssh_key_value and ssh_key_path:
+    if ssh_key_values and ssh_key_path:
         os_profile['linuxConfiguration'] = {
             'disablePasswordAuthentication': authentication_type == 'ssh',
             'ssh': {
@@ -718,7 +719,7 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
                     {
                         'path': ssh_key_path,
                         'keyData': ssh_key_value
-                    }
+                    } for ssh_key_value in ssh_key_values
                 ]
             }
         }