+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,11 +2,14 @@
 
 Release History
 ===============
-
+* functionapp: deprecated `az functionapp devops-build` command will be removed in next release
+* functionapp: `az functionapp devops-pipeline` fetches building log from Azure DevOps in verbose mode
+* functionapp: `az functionapp devops-pipeline` removes `--use_local_settings` flag, since it does not change Azure DevOps release definition
 * webapp: webapp up returns a json output if --logs is not used
 * webapp: webapp up supports writing the default resources to local configuration
 * webapp: webapp up supports redeploying an app without using the --location argument
 * webapp: Fixed an issue where for Linux Free SKU ASP creation use Free as SKU value was not working
+* Upgrade azure-mgmt-storage from 3.1.1 to 3.3.0
 
 0.2.19
 ++++++